Refactored VGG16 test to be modular and object oriented, may have an issue with train/test
